typecasting float

+++ compressai/datasets/image.py
@@ -200,7 +200,7 @@ class FeatureFolderScale(Dataset):
         t = torch.as_tensor(np.load(self.samples[index], allow_pickle=True).astype('float'))
        # t = from_numpy(np.load(self.samples[index], allow_pickle=True))
         if t.shape[2] == 256:
-            return t
+            return t.float()
         if 64 < max(t.shape[2], t.shape[3]) <= 128:     # p3
             t = interpolate(t, scale_factor=2, mode='bicubic')
         elif 32 < max(t.shape[2], t.shape[3]) <= 64:    # p4
@@ -211,7 +211,7 @@ class FeatureFolderScale(Dataset):
         hpad, wpad = 256-t.shape[2], 256-t.shape[3]
         padding = torch.nn.ZeroPad2d((math.ceil(wpad/2),math.floor(wpad/2), math.ceil(hpad/2), math.floor(hpad/2)))
         
-        return padding(t)
+        return padding(t).float()
         #print("x_hat: {}".format(x_hat[0, 1, 0, 0]))
 
         #return from_numpy(load(self.samples[index]))
